pyproject.toml
[tool.poetry]
name = "langchain-experiments"
version = "0.1.0"
description = ""
authors = ["Nicolás Marino <nicolasmarino.17@gmail.com>"]
license = "MIT"
packages = [
    { include = "langchain" }, 
]

[tool.poetry.dependencies]
python = "^3.8.1, <3.10"
openai = "^0.27.4"
langchain = "^0.0.142"
load-dotenv = "^0.1.0"
unstructured = "^0.5.12"
tiktoken = "^0.3.3"
rich = "^13.3.4"
transformers = "^4.28.1"
pypdf = "^3.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

